backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.3

# Redis & Celery
redis[hiredis]==5.1.0
celery==5.4.0

# Firebase Admin
firebase-admin==6.5.0

# AWS S3
boto3==1.35.0

# AI / ML
openai==1.50.2
transformers==4.45.0
torch==2.4.1
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.2
httpx==0.27.2

# HTTP client for weather API
aiohttp==3.10.5

# Config & validation
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
aiosqlite==0.20.0
